Initial server management refactor
