+++ source/core/server/foundry.py
@@ -1,0 +1,1 @@
+# Eventually move all the server management functions from constants to here
